+++ core/dicehub/image/client/client.go
@@ -6,7 +6,7 @@ package client
 import (
 	context "context"
 	grpc "github.com/erda-project/erda-infra/pkg/transport/grpc"
-	pb "github.com/erda-project/erda-proto-go/dicehub/image/pb"
+	pb "github.com/erda-project/erda-proto-go/core/dicehub/image/pb"
 	grpc1 "google.golang.org/grpc"
 )
